Check for retakes, ART as JSPsych

+++ jspsych_ravens/write_data.php
@@ -5,7 +5,12 @@ $post_data = json_decode(file_get_contents('php://input'), true);
 $name = "data/".$post_data['filename'].".csv"; 
 $data = $post_data['filedata'];
 // write the file to disk
-file_put_contents($name, $data);
+if (!file_exists($name)){
+	file_put_contents($name, $data);
+} else {
+	$name = 'data/r'.substr($name, 5, strlen($name)-9,).'_'.date('z-G-i').".csv";
+	file_put_contents($name, $data);
+}
 
 header('Content-Type: application/json');
 echo json_encode(['location'=>'https://www.psycholinguistics.ml/get_next_jspsych.php']);
